+++ types/remote_visitor_data_filter.go
@@ -1,9 +1,14 @@
 package types
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/Kameleoon/client-go/v3/logging"
+	"github.com/Kameleoon/client-go/v3/utils"
+)
 
 type RemoteVisitorDataFilter struct {
-	PreviousVisitAmount int
+	PreviousVisitAmount int // The value must be between 1 and 25.
 	CurrentVisit        bool
 	CustomData          bool
 	PageViews           bool
@@ -14,6 +19,7 @@ type RemoteVisitorDataFilter struct {
 	Conversion          bool
 	Experiments         bool
 	Kcs                 bool
+	VisitorCode         interface{} // Type: `bool`. Defaults to `true` if not set.
 }
 
 func DefaultRemoteVisitorDataFilter() RemoteVisitorDataFilter {
@@ -21,13 +27,23 @@ func DefaultRemoteVisitorDataFilter() RemoteVisitorDataFilter {
 		PreviousVisitAmount: 1,
 		CurrentVisit:        true,
 		CustomData:          true,
+		VisitorCode:         true,
+	}
+}
+
+func (r *RemoteVisitorDataFilter) ApplyDefaultValues() {
+	var ok bool
+	if r.VisitorCode, ok = utils.Deopt[bool](r.VisitorCode, true); !ok {
+		logging.Warning(
+			"Failed to deopt RemoteVisitorDataFilter.VisitorCode. Expected bool or nil value, got '%s'", r.VisitorCode,
+		)
 	}
 }
 
 func (r RemoteVisitorDataFilter) String() string {
 	return fmt.Sprintf(
-		"RemoteVisitorDataFilter{PreviousVisitAmount:%d,CurrentVisit:%t,CustomData:%t,PageViews:%t,"+
-			"Geolocation:%t,Device:%t,Browser:%t,OperatingSystem:%t,Conversions:%t,Experiments:%t,Kcs:%t}",
+		"RemoteVisitorDataFilter{PreviousVisitAmount:%d,CurrentVisit:%t,CustomData:%t,PageViews:%t,Geolocation:%t,"+
+			"Device:%t,Browser:%t,OperatingSystem:%t,Conversions:%t,Experiments:%t,Kcs:%t,VisitorCode:%t}",
 		r.PreviousVisitAmount,
 		r.CurrentVisit,
 		r.CustomData,
@@ -39,5 +55,6 @@ func (r RemoteVisitorDataFilter) String() string {
 		r.Conversion,
 		r.Experiments,
 		r.Kcs,
+		r.VisitorCode,
 	)
 }