+++ utils/deopt.go
@@ -1,0 +1,8 @@
+package utils
+
+func Deopt[T any](field interface{}, defaultValue T) (T, bool) {
+	if v, ok := field.(T); ok {
+		return v, true
+	}
+	return defaultValue, field == nil
+}